Add private header for MSAICrashDataProvider
