+++ Classes/MSAIEnvelopeManager.m
@@ -6,7 +6,7 @@
 #import "MSAITelemetryContext.h"
 #import "MSAITelemetryContextPrivate.h"
 #import "MSAIHelper.h"
-#import "MSAICrashDataProvider.h"
+#import "MSAICrashDataProviderPrivate.h"
 #import "MSAIHelper.h"
 
 static NSInteger const schemaVersion = 2;
